Renamed main.jsx to homepage.jsx
